FEATURE: Allow SENTRY_RELEASE as environment variable and add more data for debugging
This will add the possibility to set a release version via environment variable or via Settings.yaml.
If the env variable SENTRY_RELEASE is not set, nor the key PunktDe.Sentry.Flow.release defined with an string, it will look for
a file in the root directory which is starting with RELEASE_ and takes everything after RELEASE_ for as value for the release tag.

This fixes also a bug with Sentry Environments which won't get correctly generated by using the ->setEnvironment() function.

Some code of this is PR is heavily inspired by the flownative/flow-sentry package.